Move toggle SearchBar as SearchPanel to Header
