+++ src/components/Header/Header.module.css
@@ -1,47 +1,47 @@
+.headerSectionContainer {
+    width: 100%;
+    position: relative;
+    /* top: 0; */
+    z-index: 1000;
+}
+
 .header {
     width: 100%;
-    height: 80px;
-    z-index: 1000;
     top: 0;
-    left: 0;
     display: flex;
     justify-content: space-between;
-    align-items: center;
-    padding: 0 20px;
+    align-items: top;
+    padding: 10px 30px;
     box-sizing: border-box;
     position: fixed;
     background-color: #ffffff;
+    height: 160px;
 }
 
-.logoDiv {
-    flex: 0 0 auto;
+.separator {
+    width: 85%;
+    height: 1px;
+    background-color: #121212;
+    margin-top: 165px;
+    z-index: 1000;
 }
 
 .airbnbHomeButton {
+    display: flex;
+    align-items: top;
     cursor: pointer;
-    height: 80px;
-    width: 80px;
 }
 
 .logoAirbnb {
-    height: 80px;
+    height: 60px;
     width: 80px;
 }
 
 .beAHostAndLanguageSelectorWrapper {
     display: flex;
-    align-items: center;
-}
-
-.hostButtonDiv {
-    flex: 0 0 auto;
-    display: flex;
-    align-items: center;
-}
-
-.languageSelectorDiv {
-    display: flex;
-    align-items: center;
+    align-items: top;
+    margin-top: 10px;
+    ;
 }
 
 @media (max-width: 1024px) {
@@ -60,9 +60,4 @@
     .toggleButtonsStaysExperiences {
         display: none;
     }
-}
-
-.headerContainer {
-    width: 100%;
-    position: relative;
 }
