+++ src/components/SearchBar/SearchBar.module.css
@@ -2,12 +2,13 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  background-color: #ffffff;
   padding: 10px;
   border-radius: 50px;
   box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
   width: 850px;
-  position: relative;
+  position: fixed;
+  left: 50%;
+  transform: translate(-50%);
 }
 
 .inputContainerWhere {
@@ -28,8 +29,8 @@
   flex-direction: column;
   align-items: flex-start;
   position: relative;
-  width: 100px;
-  padding-left: 0px;
+  width: 140px;
+  padding-left: 25px;
 }
 
 .inputContainerWho {
@@ -37,25 +38,26 @@
   flex-direction: column;
   align-items: flex-start;
   position: relative;
-  width: 160px;
-  padding-left: 0px;
+  width: auto;
+  padding-left: 25px;
+  flex: 1;
 }
 
 .inputContainerDate {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  width: 231px;
-  padding-left: 0;
+  width: 281px;
+  padding-left: 25px;
 }
 
 .label {
-  margin-left: 10px;
   font-size: 12px;
   color: #666;
   font-weight: bold;
   text-align: left;
   border: none;
+  width: 100%;
 }
 
 .searchBar input {
@@ -63,17 +65,16 @@
   outline: none;
   width: 100%;
   background-color: white;
-  text-indent: 10px;
   padding: 0px 0px;
 }
 
 .separator {
-  width: 0.5px;
+  width: 1px;
   height: 30px;
   background-color: #b5b5b5f5;
   opacity: 0.5;
   margin: 0;
-  padding: 0;
+  padding: 4.5;
 }
 
 .circleButton {
